@@ -14,24 +14,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.2",
-    "@angular/common": "^5.0.2",
-    "@angular/compiler": "^5.0.2",
-    "@angular/core": "^5.0.2",
-    "@angular/forms": "^5.0.2",
-    "@angular/http": "^5.0.2",
-    "@angular/platform-browser": "^5.0.2",
-    "@angular/platform-browser-dynamic": "^5.0.2",
-    "@angular/platform-server": "^5.0.2",
-    "@angular/router": "^5.0.2",
+    "@angular/animations": "^5.2.10",
+    "@angular/common": "^5.2.10",
+    "@angular/compiler": "^5.2.10",
+    "@angular/core": "^5.2.10",
+    "@angular/forms": "^5.2.10",
+    "@angular/http": "^5.2.10",
+    "@angular/platform-browser": "^5.2.10",
+    "@angular/platform-browser-dynamic": "^5.2.10",
+    "@angular/platform-server": "^5.2.10",
+    "@angular/router": "^5.2.10",
     "angular2-recaptcha": "^0.4.1",
     "core-js": "^2.5.1",
-    "rxjs": "^5.5.2",
+    "rxjs": "^5.5.10",
     "zone.js": "^0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "^1.5.3",
-    "@angular/compiler-cli": "^5.0.2",
+    "@angular/cli": "^1.7.4",
+    "@angular/compiler-cli": "^5.2.10",
     "@angular/language-service": "^4.2.4",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
@@ -48,6 +48,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
     "tslint": "~5.3.2",
-    "typescript": "2.4.2"
+    "typescript": "^2.4.2"
   }
 }
